ai_chess_king/ai_chess_king/requirements.txt
torch>=2.1.0
numpy>=1.24.0
pygame>=2.5.0
tqdm>=4.65.0
typing_extensions>=4.5.0
python-chess>=1.9.0